+++ h3_dart/lib/src/h3_factory.base.dart
@@ -4,7 +4,7 @@ import 'package:h3_common/h3_common.dart';
 abstract class BaseH3Factory {
   /// Loads H3 from process, H3 C library must be loaded into this process first.
   /// ```dart
-  /// DynamicLibrary.open('../h3_ffi/c/h3lib/build/libh3lib.lib');
+  /// DynamicLibrary.open('../h3_ffi/c/h3lib/build/h3.so');
   /// final h3 = h3Factory.process();
   /// h3.degsToRads(123);
   /// ```
@@ -14,7 +14,7 @@ abstract class BaseH3Factory {
 
   /// Loads H3 using specified [libraryPath]
   /// ```dart
-  /// final h3 = h3Factory.byPath('../h3_ffi/c/h3lib/build/libh3lib.lib');
+  /// final h3 = h3Factory.byPath('../h3_ffi/c/h3lib/build/h3.so');
   /// h3.degsToRads(123);
   /// ```
   H3 byPath(String libraryPath);