+++ h3_dart/test/h3_dart_web_test.dart
@@ -24,7 +24,7 @@ void main() async {
       reason: 'H3Factory.process throws error',
     );
     expect(
-      () => h3Factory.byPath('../h3_ffi/c/h3lib/build/libh3lib.lib'),
+      () => h3Factory.byPath('../h3_ffi/c/h3lib/build/h3.so'),
       throwsA(isA<UnsupportedError>()),
       reason: 'H3Factory.byPath throws error',
     );