+++ h3_ffi/test/h3_ffi_factory_test.dart
@@ -7,14 +7,14 @@ void main() {
   test('H3FfiFactory', () async {
     expect(
       H3FfiFactory().byDynamicLibary(
-        DynamicLibrary.open(p.canonicalize('c/h3lib/build/libh3lib.lib')),
+        DynamicLibrary.open(p.canonicalize('c/h3lib/build/h3.so')),
       ),
       isA<H3Ffi>(),
       reason: 'H3FfiFactory.byDynamicLibary returns H3Ffi instance',
     );
 
     expect(
-      H3FfiFactory().byPath('c/h3lib/build/libh3lib.lib'),
+      H3FfiFactory().byPath('c/h3lib/build/h3.so'),
       isA<H3Ffi>(),
       reason: 'H3FfiFactory.byPath returns H3Ffi instance',
     );