+++ scripts/prepare_tests.sh
@@ -1,4 +1,4 @@
 mkdir h3_ffi/c/h3lib/build
 cd h3_ffi/c/h3lib/build
-cmake .. -G "Unix Makefiles" 
+cmake .. -G "Unix Makefiles" -DBuildForTests=ON
 make