+++ h3_ffi/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.6.2
 * Flutter for Windows support
+* Rename `libh3lib.lib` file to `h3.so`
 
 ## 0.6.1
 * Add holes support to `polyfill` function (thanks [@iulian0512](https://github.com/iulian0512))