+++ h3_flutter/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.6.4
 * Flutter for Windows support
+* Rename `libh3lib.lib` file to `h3.so`
 
 ## 0.6.3
 * Add `holes` param to `H3.polyfill` function (thanks [@iulian0512](https://github.com/iulian0512))